Added IBookRepository to HomeController.  Updated the BookRepository to map DigitalBook.

+++ DigitalBookHistoryApp/API/DigitalBookHistoryAPI/Models/DigitalBook.cs
@@ -13,8 +13,8 @@ namespace DigitalBookHistoryAPI.Models
         public int KindId { get; set; }
         public string ArtistName { get; set; } // TODO: Refactor into List<Artists>?
         public string ArtKey { get; set; }
-        public int Borrowed { get; set; }
-        public int Returned { get; set; }
+        public Int64 Borrowed { get; set; }
+        public Int64 Returned { get; set; }
 
         public Image Image { get; set; }
         public Kind Kind { get; set; }
